PUT method

+++ src/main/java/edu/infsci2560/controllers/LocationsController.java
@@ -24,6 +24,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.servlet.ModelAndView;
 
@@ -53,12 +54,15 @@ public class LocationsController {
     }
     
 
-    @RequestMapping(value = "locations/delete", method = RequestMethod.DELETE)
-    @ResponseStatus(HttpStatus.NO_CONTENT)
-	public void delete(@PathVariable("id") long id) {
-    	repository.delete(id);
+    @RequestMapping(value = "locations/delete", method = RequestMethod.GET)
+    public ModelAndView delete(@RequestParam(value="id", required=true) Long id) {
+        Location location = repository.findOne(id);  
+        if ( location != null ) {
+            repository.delete(id);
+        }
+        return new ModelAndView("locations", "locations", repository.findAll());
+    }
 
-	}
     
     
 }
